MH-12502, Convert Workspace Read To InputStream
This patch makes the `workspace.read()` method return an InputStream,
ensuring that the files content which may be read directly from the
working file repository (if possible) is not modified.

In case the files cannot be read directly (e.g. non-local files), the
files are temporarily copied into the workspace, but will automatically
be removed once the returned InputStream is closed to not leave any
files lying around in the workspace.