MH-13227, Update to Apache Karaf 4.2
This patch updates Opencast to Apache Karaf 4.2 which is the latest
stable branch of Karaf right now. It updates a number of libraries in
the process with the ActiveMQ client library being the most notably
one.

ActiveMQ was hardened against exploitation of Java object
(de-)serialization in the meantime. Something Opencast utterly relies on
for messaging. This patch now allows Opencast's messages explicitly,
preventing any other objects from being sent.