MH-12091, Per-Tenant Capture Agent Users
This patch re-implements the feature of per-tenant capture agent users
which can be used with HTTP Digest authentication.

It provides a configuration file in which users can be defined and be
assigned to organizations. Additional roles can be applied as well.

This patch also allows these users access to the asset manager so that
files ingested by a capture agent can be archived as part of the
launched workflow.