MH-13320, Get ACL Rules From Database
Instead of building an SQL query including all of a users roles to check
against the access rules stored in the asset manager's database table
which can result in a very large and complex query, this patch allows us
to turn the check around, extract the usually few rules from the
database in an efficient manner and then check the users roles against
those.

Unfortunately, this is only possible with some more specific queries
though. The general query DSL requests still reply on the in-database
checks.