Fast studio workflow and encoding profiles created
The default opencast workflows have some issues processing video files created with opencast studio. As we also want an simplified workflow to process these files as fast as possible, I've created a new workflow to cover this case. The encoding profiles are optimized for performance and fixes the issues with uncommon video resolutions and variable frame rate.