Disable dependabot for this repository
This is a fork that will regularly rebased on top of the community
code. Having duplicate dependabot PRs here does not make sense.

As described in [1]: "You can disable version updates entirely by
deleting the `dependabot.yml` file from your repository." But there
also dependabot security updates which are enabled even without
`dependabot.yml`. But those are disabled for forks as described in [2].

[1]: https://docs.github.com/en/code-security/supply-chain-security/keeping-your-dependencies-updated-automatically/enabling-and-disabling-version-updates#disabling-dependabot-version-updates
[2]: https://docs.github.com/en/code-security/supply-chain-security/managing-vulnerabilities-in-your-projects-dependencies/configuring-dependabot-security-updates#supported-repositories