Clean up server status endpoint
This patch updates the code for the admin interface server status
endpoint. It tries to simplify the code, handles a number of different
user errors which previously resulted in internal server errors or
weird results and replaced the Entwine functional library.

This does not change the current functionality in any meaningful way nor
fixes it the performance problem.